+++ docs/src/pages/MainPage.tsx
@@ -23,6 +23,12 @@ export default function MainPage() {
               </Button>
 
               <Button
+                onClick={() => navigate('/avatar-demo')}
+              >
+                Avatar
+              </Button>
+
+              <Button
                 onClick={() => navigate('/box-demo')}
               >
                 Box